Add --assume-odk-repository option.
The ODK Runner normally tries to detect automatically whether it is
called from within the src/ontology directory of a ODK repository, by
(1) looking for a *-odk.yaml file and (2) checking that the last two
components of the current directory pathname are `src/ontology`. If it
finds that it is indeed running from within a ODK repository, it binds
the top-level directory of the repository to the container (two
directories above the current directory) rather than the current
directory itself.

But some ODK repositories may not contain a src/ontology/*-odk.yaml file
(Mondo, at least, does not have one). For those, the
--assume-odk-repository option may be used to force the runner to act as
if it was running from within a ODK repository.