Add new util function get_data_directory.
Move the code that finds the per-user data directory (currently used in
the `oaklib` module) into a new helper function in the `util` module,
because we are about to reuse that code elsewhere.

We use that opportunity to fix a bug in the GNU/Linux-specific portion
of that code: we were looking up for the XDG_DATA_DIR variable, instead
of XDG_DATA_HOME. This still worked in most cases because when we cannot
find XDG_DATA_DIR (which has no reason to exist; beware that this is not
XDG_DATA_DIRS), we fall back to `$HOME/.local/share`. But this would
still cause issue if the the user has set up their system so that
XDG_DATA_HOME is not in its default location.