Support odkcore-generated native ODK environments.
Amend the behavior of the `--native` option. When used, instead of
assuming that all ODK tools are in the PATH (which will not be enough
for the ODK workflows to work properly, once the split between ODK-Image
and ODK-Core will be complete), we instead assume that the name of the
ODK "image" to use (as given to the `--image` option) refers to a
directory under `$XDG_DATA_HOME/ontology-development-kit/images` that
has been initialised with `odk install`.