Refactor the code to get the per-user directories.
Move the definitions of the per-user directories for each system in a
separate usersdirs.h include file, and provide two functions to get the
path to any of the 3 possible per-user directories we currently have to
deal with (configuration directory, data directory, directory for native
environments).

The two functions are similar except that one automatically allocates a
buffer while the other expects a caller-provided buffer. The latter is
needed because the code in oaklib.c currently uses a fixed-size buffer;
that code should be reworked to use an automatically buffer as well.

Of note, the fact that we use the new function in odkrun.c to get the
Github token means that on Windows, we now look for a file called
`token`, instead of `token.txt`; that code should be amended so that we
always (on all platforms) check for `token.txt` first, then for `token`
if we could not find `token.txt`.